feat: introduce FMEA service
